feat: grade sem filter, make grades clientside
